Add cc-by-sa-4.0 licence
Open source software licenses can be used for non-software works, and often are the best choice. This is particularly the case when the works in question can be edited and versioned as source. 

CC-BY-SA-4.0 licence are recommended by [Choose a License.com](choosealicense.com/) for non-software material such as documentation.